fix(initramfs): Fix case where we download a non archive file
When the previous changes went in for the Install() func, we download to the appropriate
destination and return the sourceFile as the downloaded file. Since the sourceFile is the
actual file and not an archive, it triggers the second match in the switch statement.
Because we dont have a destFile defined, we get a fun seg fault.

Signed-off-by: Brad Beam <brad.beam@b-rad.info>