chore: use protobuf compiler from the toolchain image
This relies on new toolchain image from autonomy/toolchain-musl#9 to
provide `protoc` and `protoc-gen-go`.

New make target `proto` exports generated `.pb.go` files to the working
directory, so that Go source can is buildable which is really nice for
the editors to provide linting, to run the tests in local mode, etc.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>