chore: refactor and dry up process runner
Signficant change is going from `cmd.Process.Wait()` to `cmd.Wait()`,
the reason is that looks like `cmd.Wait()` does some stuff besides
waiting for the process to exit:

```
Wait waits for the command to exit and waits for any copying to
stdin or copying from stdout or stderr to complete.

...

The returned error is nil if the command runs, has no problems copying
stdin, stdout, and stderr, and exits with a zero exit status.

...

Wait releases any resources associated with the Cmd.
```

This replaces potentially infinite recursison with explicit loop.

I want to add unit-test for this code, but that requires adding 'Stop()'
method which requires even more refactoring, so I didn't want to go that
route so that diff is manageable. I will get unit-test done as the next
step.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>