chore: rework process runner, add tests and stop method
This is just first step towards better control of the services. Service
eventually uses either Process or Containerd runners to launch the
program. We don't have a way to control the service (stop it, figure out
whether it's actually running) unless we can pass back Runner instance
from the Service.

The idea is to return interface `runner.Runner` from the `Service` so
that controller (`init`) can figure out current state of the running
services and stop them on shutdown.

I haven't touched containerd runner yet (and as interface is not
enforced, code compiles fine). Only process runner is updated to allow
process to be stopped (either when it's running or while waiting to be
restarted).

For the sake of unit-tests, some additional runner options were exposed:
log path, restart interval, etc.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>