chore: add dev-test make target to quickly re-run unit-tests
This is a shortcut when working on changes in Go code to re-run Go tests
quickly without even trying to rebuild `test` container.

This target doesn't do any checks, it is assumed that `make test` was
run to create initial image of the `test` container. As Go sources are
mapped into the container, it picks up any changes made.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>