chore: workaround flaky tests
Some unit-tests depend on timing as they're testin concurrent code and
tests do things like `time.Sleep(...)` to let the code in concurrent
goroutine proceed so that side-effect can be observed.

One issue is that `go test` (and `go build`) by default launch builds
with the number of CPUs available which might be too much and overloads
system so that tests which are also concurrent are not allowed to
execute due to high load average. I observe "lock-ups" when running
`make test` locally. I'm trying to limit number of executors to 4 as an
experiment.

Disable short test run, as we don't seem to have any short tests, and
running short and full run just means running all the tests twice.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>