fix(init): Dont log an error when context canceled
When we receive all the necessary files from trustd, we cancel the context. This
was treated as an error case and a message was logged accordingly. However,
this case was not really an error versus a signal to stop trying to fetch a
given file.

Fixes #723

Add basic FileSet tests. Minor refactor to FileSet call to allow easier testing
Add context canceled test for download
Add config tests and trustd coverage

Signed-off-by: Brad Beam <brad.beam@talos-systems.com>