fix: Only generate pki from trustd if not control plane
This PR will fix a bug where the non-init nodes were not generating
their certs locally and relying on trustd instead. This broke down
because we aren't saving the CA key when we're generating with the
trustd identity function (because we don't need it for workers).

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>
(cherry picked from commit a15499d25aa2024187c7a9a0b53ab333cdeedced)