chore: fix generate version flag and mark v0 as deprecated
Since the command's name is 'generate' the 'gen' prefix is not needed
in the version flag. The flag is scoped under the generate command so
it should be very clear that the '--version' flag is used to control the
config version.

We also move to defaulting to v0 since v1 is new and still needs to be
tested in the real world. We can default to v1 in the next release.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>