refactor: add helper func to create cert and key
It is common for us to create a certificate and key using a root CA.
This adds a helper function to make this process easier and
standardized. For now, the function is only used when generating
configs, but there are a number of places we should update in follow
ups.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>