chore: fix flaky constant retry test
Failure:

```
--- FAIL: Test_constantRetryer_Retry (7.00s)
    --- FAIL: Test_constantRetryer_Retry/test_expected_number_of_retries (2.00s)
        constant_test.go:168: expected count of 2, got 3
```

The problem is that retry interval (1s) perfectly aligns with timeout
(2s), so depending on which timer fires first, function might be called
two or three times. Fix that by extending timeout a bit so it fits one
more run and not more.

P.S. This test might be still flaky under load if function doesn't have
a chance to run (starvation). Proper fix is to use fake time in the
tests.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>