chore: fix containerd test hanging
The problem was that if container fails to start, it never reaches
'StateRunning' and test hangs waiting for that state. Assertion doesn't
abort whole test (it only aborts goroutine it was called from), so this
doesn't help.

Fix that by signalling back if some containers fail to start.

This is not a fix, but it should expose the actual failure happening in
this test.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>