fix: conditionally create a new etcd cluster
This fixes a long standing issue with upgrading the init node. We
currently have no way of knowing whether the init node should join an
existing etcd cluster, or create a new one. This makes use of the node's
metadata to determine if the node has already created the etcd cluster.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>