fix: mount extra disks after system disk
The extra disks functionality was completely broken. One fundamental
issue was that we were attempting to create and mount the partitions
before the system disk was created. This moves the extra disks tasks to
the correct part of the boot sequnce. This also adds a simple check that
refuses to operate on a disk if any partitions are found.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>