feat: Add support for kernel network config (ip=)
This introduces a kernel addressing method. Given that the kernel handles this configuration
this mode is largely noop ( address should already exist/be defined ), however, this adds
the ability to handle the addressing method as a first class citizen. TLDR; we can support
configuring resolv.conf via kernel ip= argument.

Signed-off-by: Brad Beam <brad.beam@talos-systems.com>