fix: ensure etcd comes back up on reboot of all members
This adds a check for the existence of data in the etcd data directory,
that we then use to conditionally set the initial cluster args to handle
the case of a reboot of all etcd members at once. Without this, etcd
fails to come back up, and effectively kills the cluster.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>