fix: error reporting in `osctl kubeconfig`
Problem seems to be on multiple levels, and there are a bit of changes
which got mixed in from another PR (just same file changed).

Core of the issue is that `helpers.Fatalf()` calls `os.Exit()` which
terminates execution and doesn't let the `defer` and other handlers to
run. This uses Cobra feature of error propagation to pop errors through
the stack back to root command.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>