fix: don't use netrc
The getter package tries to use the netrc file to gather credentials if
none are provided when using http and https getters. This disables that
because under the hood, it uses another package (go-homedir) to try and
find the netrc file. The issue is then that go-homedir tries to invoke
getent, but that is not installed on the rootfs so the command fails.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>