feat: prepend cluster name to config generate
This PR will take the cluster name and add it to the node type that is
generated by osctl config generate (ex: "mycluster-init.yaml"). Driver
for this is to help make it clearer which cluster a given yaml applies
to. Will also help with the ux of being able to config docker-based
talos clusters.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>