chore: fix push events
Now that we have a push target and push-% target, we can simplify the drone
conditions. This updates the conditions so that the latest channel updates
on pushes to master, the edge channel updates on successful nightly cron, and
an image with the standard tag is pushed in all events except pull requests.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>