fix(networkd): Fix incorrect resolver settings
This modifies the way the hostname gets set. Previously, we would run
through the entire addressing and resolver configuration and then set the
hostname. This is problematic because the resolver depends on the functionality
of Hostname() ( resolver configuration relies on the domainname of the host ).

Signed-off-by: Brad Beam <brad.beam@talos-systems.com>
(cherry picked from commit 93218687ec1a2a3116911d66a45d200461b02b02)