chore: validate installer image before upgrade
This PR will add a very basic check to make sure that the installer
image specified during something like `osctl upgrade` indeed seems to be
a talos installer. We do this by simply calling the help flag of a known
binary (`/bin/installer --help`). If this doesn't exist, we exit with
the assumption that the user has made a mistake.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>