fix: don't use ARP table for networkd health check
We saw strange behavior in this health check, and think that it is due
to ARP table cache. In practice, the health check caused nodes to hang.
We decided to not use the ARP table to determine the health of networkd.
The fact that networkd can respond to the health API should be
sufficient enough since network will fail to start upon any error in the
initial setup of the network.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>