fix: wait for USB storage
This fixes a race condition between config validation and USB storage
availablity. If a user attempts to install to a USB disk, we should wait
for the amount of seconds specified in
`/sys/module/usb_storage/parameters/delay_use` before we attempt to
validate the config. If we don't, then there is the chance that the USB
device is not mounted in time, and we fail to install because we cannot
verify the disk is available.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>