refactor: rename system-containerd and containerd services
This changes the name of the `system-containerd` service to `containerd`
and the `containerd` service to `cri`. This is so that the CRI service
is a little more generic. In the future, when we add support for other
CRIs, it will be better to refer to this service generically as "cri."

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>