chore: use a single CHANGELOG
Instead of keeping a CHANGELOG for each release in the master branch, a
single CHANGELOG should be used since it will move into release branches
anyways. This prevents us from having to keep the files in sync across
master and the release branch. This also adds better tooling for
generating the CHANGELOG.md.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>