fix: prevent formatting the ephemeral partition twice
This adds the requirement of the existence of the boot partition in the
if statement responsible for handling v0.4 style upgrades. Without this,
when upgrading from v0.3 to v0.4, we will attempt to format the
ephemeral partition twice. This will cause upgrades to fail.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>
(cherry picked from commit f77c353e657d8204c0f4769b0841ce1246725b13)